// quickapp.client/src/app/models/AppTheme.ts
// ======================================
// Author: Ebenezer Monney
// Copyright (c) 2023 www.ebenmonney.com
// 
// ==> Gun4Hire: contact@ebenmonney.com
// ======================================

export interface AppTheme {
  id: number;
  name: string;
  displayName: string;
  primary: string;
  secondary: string;
  isDefault?: boolean;
}